Implemented the first web-service
Implemented the first web-service and web-service client. Used apache cxf
maven plugin to generate client sources from wsdl and
build-helper-maven-plugin to include the source folder. The web-service
and wildfly needs to be running of course, if you want to generate sources
from wsdl (otherwise the build will fail of course :) )